requirements.txt
spotipy
python-twitter